improving example
re-writing the smarteole example

- made the notebook more clear and included steps previously hidden in .py file
- removing code only used in testing from example files which made test_smarteole.py more self-standing
- replace test_data with zip file from zenodo (which is smaller) and used as cache in both tests and example ipynb

Wanted to move dev to extras:
- Extras = define a set of dependencies, that will bring extra functionality to your package
- Dependency Groups = storing package requirements that are not included in project metadata
however cannot use dep groups until `extra` until pip 25.1
ref: https://github.com/pypa/pip/pull/13065#issuecomment-2675797688
